requirements.txt
altgraph==0.17.4
importlib-metadata==6.8.0
packaging==23.2
passlib==1.7.4
pkg_resources==0.0.0
pycryptodome==3.19.0
pycryptodomex==3.19.0
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
zipp==3.17.0
